+++ CHANGELOG.md
@@ -29,3 +29,4 @@
 - [feat: support otel trace](https://github.com/Tencent/spring-cloud-tencent/pull/1363)
 - [feat:support TSF router.](https://github.com/Tencent/spring-cloud-tencent/pull/1368)
 - [feat:upgrade nearby router and add namespace nearby router.](https://github.com/Tencent/spring-cloud-tencent/pull/1371)
+- [fix:fix contract reporting bug when using TSF.](https://github.com/Tencent/spring-cloud-tencent/pull/1373)